Fix by adding error message while add to cart (anonymous)

+++ src/actions/cart.actions.ts
@@ -9,6 +9,14 @@ export async function getUserCart() {
     try {
 
         const token = await getUserToken()
+        if (!token) {
+            return {
+                error: {
+                    message: "Please login first"
+                },
+                status: false
+            }
+        }
 
         const resp = await fetch(`https://ecommerce.routemisr.com/api/v2/cart`, {
             method: "GET",
@@ -45,6 +53,14 @@ export async function addToCart(productId: string) {
     try {
 
         const token = await getUserToken()
+        if (!token) {
+            return {
+                error: {
+                    message: "Please login first"
+                },
+                status: false
+            }
+        }
 
         const resp = await fetch(`https://ecommerce.routemisr.com/api/v2/cart`, {
             method: "POST",
@@ -82,6 +98,14 @@ export async function updateProductQty(productId: string, count: number) {
     try {
 
         const token = await getUserToken()
+        if (!token) {
+            return {
+                error: {
+                    message: "Please login first"
+                },
+                status: false
+            }
+        }
 
         const resp = await fetch(`https://ecommerce.routemisr.com/api/v2/cart/${productId}`, {
             method: "PUT",
@@ -124,6 +148,14 @@ export async function removeProduct(productId: string) {
     try {
 
         const token = await getUserToken()
+        if (!token) {
+            return {
+                error: {
+                    message: "Please login first"
+                },
+                status: false
+            }
+        }
 
         const resp = await fetch(`https://ecommerce.routemisr.com/api/v2/cart/${productId}`, {
             method: "DELETE",
@@ -166,6 +198,14 @@ export async function clearUserCart() {
 
         const token = await getUserToken()
 
+        if (!token) {
+            return {
+                error: {
+                    message: "Please login first"
+                },
+                status: false
+            }
+        }
         const resp = await fetch(`https://ecommerce.routemisr.com/api/v2/cart`, {
             method: "DELETE",
             headers: {
